Interactive Table of Descriptive Statistics in HTML

Interactive Table of Descriptive Statistics in HTML

Usage

metalite_table1(
  formula,
  data,
  id,
  var_listing = NULL,
  total = TRUE,
  download = "none",
  type = NULL
)

Arguments

formula

an object of class "formula".

data

a data frame that contain variables described in the formula.

id

a character value to indicate subject/record id variable name in data.

var_listing

a character vector of additional variables included in the drill down listing.

total

a logical value to display or hide "Total" column.

download

a character value to enable download button. Allowed values include "none", "listing", "table", and 'all'.

type

a character value to control section title (e.g. "Subjects", "Records").

Examples

if(interactive()){
  metalite_table1(~ AGE + SEX | TRT01A, data = r2rtf::r2rtf_adsl, id = "SUBJID")
}